Adding switch for firewall option and prerequisites. Closing #25

+++ helpers/tcp.psm1
@@ -8,24 +8,15 @@ function Invoke-TCP {
         [PSObject] $Sender
     )
 
-    $TCPResults = New-Object -TypeName psobject
+    $ModuleBase = (Get-Module Test-NetStack -ListAvailable | Select-Object -First 1).ModuleBase
 
-    if (!($Receiver.IPAddress -in $global:localIPs) -and !(Invoke-Command -ComputerName $Receiver.NodeName -ScriptBlock { Test-Path -Path "C:\Test-NetStack\tools\CTS-Traffic\ctsTraffic.exe" })) {
-        $DestinationSession = New-PSSession -ComputerName $Receiver.NodeName
-        Invoke-Command -ComputerName $Receiver.NodeName -ScriptBlock {cmd /c "mkdir C:\Test-NetStack\tools\CTS-Traffic"} -ErrorAction SilentlyContinue
-        Copy-Item C:\Test-NetStack\tools\CTS-Traffic\ctsTraffic.exe -Destination C:\Test-NetStack\tools\CTS-Traffic -Force -ToSession $DestinationSession -ErrorAction SilentlyContinue
+    if ($EnableFirewallRules) {
+        Invoke-Command -ComputerName $Receiver.NodeName, $Sender.NodeName -ScriptBlock { New-NetFirewallRule -DisplayName "Client-To-Server Network Test Tool" -Direction Inbound -Program "$ModuleBase\tools\CTS-Traffic\ctsTraffic.exe" -Action Allow | Out-Null }
     }
-    if (!($Sender.IPAddress -in $global:localIPs) -and !(Invoke-Command -ComputerName $Sender.NodeName -ScriptBlock { Test-Path -Path "C:\Test-NetStack\tools\CTS-Traffic\ctsTraffic.exe" })) {
-        $DestinationSession = New-PSSession -ComputerName $Sender.NodeName
-        Invoke-Command -ComputerName $Sender.NodeName -ScriptBlock {cmd /c "mkdir C:\Test-NetStack\tools\CTS-Traffic"} -ErrorAction SilentlyContinue
-        Copy-Item C:\Test-NetStack\tools\CTS-Traffic\ctsTraffic.exe -Destination C:\Test-NetStack\tools\CTS-Traffic -Force -ToSession $DestinationSession -ErrorAction SilentlyContinue
-    }
-    
-    Invoke-Command -ComputerName $Receiver.NodeName, $Sender.NodeName -ScriptBlock { New-NetFirewallRule -DisplayName "Client-To-Server Network Test Tool" -Direction Inbound -Program "C:\Test-NetStack\tools\CTS-Traffic\ctsTraffic.exe" -Action Allow | Out-Null }
 
     # CTS Traffic Rate Limit is specified in bytes/second
     $ServerLinkSpeed = $Receiver.LinkSpeed.split(" ")
-    Switch($ServerLinkSpeed[1]) {            
+    Switch($ServerLinkSpeed[1]) {
         ("Gbps") {$ServerLinkSpeedBps = [Int]::Parse($ServerLinkSpeed[0]) * [Math]::Pow(10, 9) / 8}
         ("Mbps") {$ServerLinkSpeedBps = [Int]::Parse($ServerLinkSpeed[0]) * [Math]::Pow(10, 6) / 8}
         ("Kbps") {$ServerLinkSpeedBps = [Int]::Parse($ServerLinkSpeed[0]) * [Math]::Pow(10, 3) / 8}
@@ -33,92 +24,93 @@ function Invoke-TCP {
     }
 
     $ClientLinkSpeed = $Sender.LinkSpeed.split(" ")
-    Switch($ClientLinkSpeed[1]) {              
+    Switch($ClientLinkSpeed[1]) {
         ("Gbps") {$ClientLinkSpeedBps = [Int]::Parse($ClientLinkSpeed[0]) * [Math]::Pow(10, 9) / 8}
         ("Mbps") {$ClientLinkSpeedBps = [Int]::Parse($ClientLinkSpeed[0]) * [Math]::Pow(10, 6) / 8}
         ("Kbps") {$ClientLinkSpeedBps = [Int]::Parse($ClientLinkSpeed[0]) * [Math]::Pow(10, 3) / 8}
         ("bps") {$ClientLinkSpeedBps = [Int]::Parse($ClientLinkSpeed[0]) / 8}
     }
-    
+
     $ServerRecvCounter = Start-Job `
     -ScriptBlock {
         param([string]$ServerName,[string]$ServerInterfaceDescription)
-        $ServerInterfaceDescription = (((($ServerInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_' 
+        $ServerInterfaceDescription = (((($ServerInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_'
         Invoke-Command -ComputerName $ServerName `
-        -ScriptBlock { 
+        -ScriptBlock {
             param([string]$ServerInterfaceDescription)
-            Get-Counter -Counter "\Network Adapter($ServerInterfaceDescription)\Bytes Received/sec" -MaxSamples 20 -ErrorAction Ignore 
+            Get-Counter -Counter "\Network Adapter($ServerInterfaceDescription)\Bytes Received/sec" -MaxSamples 20 -ErrorAction Ignore
          } `
          -ArgumentList $ServerInterfaceDescription
     } `
     -ArgumentList $Receiver.NodeName,$Receiver.InterfaceDescription
-    
+
     $ServerSendCounter = Start-Job `
     -ScriptBlock {
         param([string]$ServerName,[string]$ServerInterfaceDescription)
-        $ServerInterfaceDescription = (((($ServerInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_' 
+        $ServerInterfaceDescription = (((($ServerInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_'
         Invoke-Command -ComputerName $ServerName `
-        -ScriptBlock { 
+        -ScriptBlock {
             param([string]$ServerInterfaceDescription)
-            Get-Counter -Counter "\Network Adapter($ServerInterfaceDescription)\Bytes Sent/sec" -MaxSamples 20 -ErrorAction Ignore 
+            Get-Counter -Counter "\Network Adapter($ServerInterfaceDescription)\Bytes Sent/sec" -MaxSamples 20 -ErrorAction Ignore
          } `
          -ArgumentList $ServerInterfaceDescription
     } `
     -ArgumentList $Receiver.NodeName,$Receiver.InterfaceDescription
 
-    $ServerOutput = Start-Job `
-    -ScriptBlock {
-        param([string]$ServerName,[string]$ServerIP,[string]$ServerLinkSpeed)
-        Invoke-Command -ComputerName $ServerName `
-        -ScriptBlock { 
-            param([string]$ServerIP)
-            cmd /c "C:\Test-NetStack\tools\CTS-Traffic\ctsTraffic.exe -listen:$ServerIP -consoleverbosity:1 -ServerExitLimit:64 -TimeLimit:20000 -pattern:duplex" 
-         } `
-         -ArgumentList $ServerIP
-    } `
-    -ArgumentList $Receiver.NodeName,$Receiver.IPAddress,$Receiver.LinkSpeed
+    $ServerOutput = Start-Job -ScriptBlock {
+        param([string] $ServerName, [string] $ServerIP, $ModuleBase)
 
-    $ClientRecvCounter = Start-Job `
-    -ScriptBlock {
-        param([string]$ClientName,[string]$ClientInterfaceDescription) 
-        $ClientInterfaceDescription = (((($ClientInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_' 
-        Invoke-Command -ComputerName $ClientName `
-        -ScriptBlock { 
-            param([string]$ClientInterfaceDescription)
-            Get-Counter -Counter "\Network Adapter($ClientInterfaceDescription)\Bytes Received/sec" -MaxSamples 20 
-         } `
-         -ArgumentList $ClientInterfaceDescription
-    } `
-    -ArgumentList $Sender.NodeName,$Sender.InterfaceDescription
+        Invoke-Command -ComputerName $ServerName -ScriptBlock {
+            param(
+                [string] $ServerIP,
+                [string] $ModuleBase
+            )
 
-    $ClientSendCounter = Start-Job `
-    -ScriptBlock {
-        param([string]$ClientName,[string]$ClientInterfaceDescription)
-        $ClientInterfaceDescription = (((($ClientInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_' 
-        Invoke-Command -ComputerName $ClientName `
-        -ScriptBlock { 
+            & "$ModuleBase\tools\CTS-Traffic\ctsTraffic.exe -listen:$ServerIP -Protocol:tcp -buffer:262144 -transfer:21474836480 -Pattern:push -TimeLimit:30000"
+         } -ArgumentList $ServerIP, $ModuleBase
+
+    } -ArgumentList $Receiver.NodeName, $Receiver.IPAddress, $ModuleBase
+
+    $ClientRecvCounter = Start-Job -ScriptBlock {
+        param([string] $ClientName, [string] $ClientInterfaceDescription)
+
+        $ClientInterfaceDescription = (((($ClientInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_'
+
+        Invoke-Command -ComputerName $ClientName -ScriptBlock {
             param([string]$ClientInterfaceDescription)
+
+            Get-Counter -Counter "\Network Adapter($ClientInterfaceDescription)\Bytes Received/sec" -MaxSamples 20
+         } -ArgumentList $ClientInterfaceDescription
+
+    } -ArgumentList $Sender.NodeName, $Sender.InterfaceDescription
+
+    $ClientSendCounter = Start-Job -ScriptBlock {
+        param([string] $ClientName, [string] $ClientInterfaceDescription)
+
+        $ClientInterfaceDescription = (((($ClientInterfaceDescription) -replace '#', '_') -replace '[(]', '[') -replace '[)]', ']') -replace '/', '_'
+
+        Invoke-Command -ComputerName $ClientName -ScriptBlock {
+            param([string] $ClientInterfaceDescription)
+
             Get-Counter -Counter "\Network Adapter($ClientInterfaceDescription)\Bytes Sent/sec" -MaxSamples 20
-         } `
-         -ArgumentList $ClientInterfaceDescription
-    } `
-    -ArgumentList $Sender.NodeName,$Sender.InterfaceDescription
-    
-    $ClientOutput = Start-Job `
-    -ScriptBlock {
-        param([string]$ClientName,[string]$ServerIP,[string]$ClientIP,[string]$ClientLinkSpeed)
-        Invoke-Command -ComputerName $ClientName `
-        -ScriptBlock { 
-            param([string]$ServerIP,[string]$ClientIP,[string]$ClientLinkSpeed)
-            cmd /c "C:\Test-NetStack\tools\CTS-Traffic\ctsTraffic.exe -target:$ServerIP -bind:$ClientIP -connections:64 -consoleverbosity:1 -pattern:duplex"  
-         } `
-         -ArgumentList $ServerIP,$ClientIP,$ClientLinkSpeed
-    } `
-    -ArgumentList $Sender.NodeName,$Receiver.IPAddress,$Sender.IPAddress,$ClientLinkSpeedBps
-    
+         } -ArgumentList $ClientInterfaceDescription
 
-    Start-Sleep 20
-                       
+    } -ArgumentList $Sender.NodeName,$Sender.InterfaceDescription
+
+    $ClientOutput = Start-Job -ScriptBlock {
+        param([string] $ClientName, [string] $ServerIP, [string] $ClientIP, [string] $ModuleBase)
+
+        Invoke-Command -ComputerName $ClientName -ScriptBlock {
+            param( [string] $ServerIP, [string] $ClientIP, $ModuleBase )
+
+            & "$ModuleBase\tools\CTS-Traffic\ctsTraffic.exe -target:$ServerIP -bind:$ClientIP -Connections:64 -Iterations:1 -Protocol:tcp -buffer:262144 -transfer:21474836480 -Pattern:push"
+         } -ArgumentList $ServerIP, $ClientIP, $ModuleBase
+
+    } -ArgumentList $Sender.NodeName, $Receiver.IPAddress, $Sender.IPAddress, $ModuleBase
+
+    #TODO: We should track the job state or listen for an event rather than a simplistic timer like this.
+    Start-Sleep 30
+
     $ServerRecv = Receive-Job $ServerRecvCounter
     $ServerSend = Receive-Job $ServerSendCounter
     $ClientRecv = Receive-Job $ClientRecvCounter
@@ -141,12 +133,12 @@ function Invoke-TCP {
     $ServerSendBitsPerSecond = [Math]::Round(($FlatServerSendOutput | Measure-Object -Maximum).Maximum, 2)
     $ClientRecvBitsPerSecond = [Math]::Round(($FlatClientRecvOutput | Measure-Object -Maximum).Maximum, 2)
     $ClientSendBitsPerSecond = [Math]::Round(($FlatClientSendOutput | Measure-Object -Maximum).Maximum, 2)
-                      
+
     Write-Verbose "Server Recv bps: $ServerRecvBitsPerSecond"
     Write-Verbose "Server Send bps: $ServerSendBitsPerSecond"
     Write-Verbose "Client Recv bps: $ClientRecvBitsPerSecond"
     Write-Verbose "Client Send bps: $ClientSendBitsPerSecond"
-                        
+
     $ServerOutput = Receive-Job $ServerOutput
     $ClientOutput = Receive-Job $ClientOutput
 
@@ -167,12 +159,15 @@ function Invoke-TCP {
     $ReceivedGbps = [Math]::Round($ServerRecvBitsPerSecond * [Math]::Pow(10, -9), 2)
     $ReceivedPercentageOfLinkSpeed = [Math]::Round(($ReceivedGbps / $ReceiverLinkSpeedGbps) * 100, 2)
 
+    $TCPResults = New-Object -TypeName psobject
     $TCPResults | Add-Member -MemberType NoteProperty -Name ReceiverLinkSpeedGbps -Value $ReceiverLinkSpeedGbps
     $TCPResults | Add-Member -MemberType NoteProperty -Name ReceivedGbps -Value $ReceivedGbps
     $TCPResults | Add-Member -MemberType NoteProperty -Name ReceivedPctgOfLinkSpeed -Value $ReceivedPercentageOfLinkSpeed
     $TCPResults | Add-Member -MemberType NoteProperty -Name RawData -Value $RawData
 
-    Invoke-Command -ComputerName $Receiver.NodeName, $Sender.NodeName -ScriptBlock { Remove-NetFirewallRule -DisplayName "Client-To-Server Network Test Tool" | Out-Null }
+    if ($EnableFirewallRules) {
+        Invoke-Command -ComputerName $Receiver.NodeName, $Sender.NodeName -ScriptBlock { Remove-NetFirewallRule -DisplayName "Client-To-Server Network Test Tool" | Out-Null }
+    }
 
     Return $TCPResults
 }
